+++ config.ini
@@ -7,5 +7,5 @@
 #
 
 [global]
-version=2.2.1
+version=2.3.0
 release=1
