+++ src/proxy/src/status_reporter.rs
@@ -1,6 +1,6 @@
 use crate::controller::ConnectionSearchState;
 use crate::efs_rpc::PartitionId;
-use crate::{proxy::PerformanceStats, proxy_identifier::ProxyIdentifier};
+use crate::{proxy_identifier::ProxyIdentifier, proxy_task::PerformanceStats};
 use anyhow::{Error, Result};
 use tokio::sync::mpsc::{self, Receiver, Sender};
 use tokio::time::Instant;
